改成 pgsql 才支援的 column

+++ webdata/models/Track.php
@@ -136,7 +136,7 @@ class Track extends Pix_Table
         $this->_columns['url'] = array('type' => 'varchar', 'size' => 255);
         $this->_columns['options'] = array('type' => 'text');
         // 0-每日, 1-每五分鐘
-        $this->_columns['track_period'] = array('type' => 'tinyint');
+        $this->_columns['track_period'] = array('type' => 'int');
     }
 
     public static function getTrackPeriods()
